Potential fix for get window position when fullscreen under X11
Some setups will combine all monitors into a single display, and as
a result, "fullscreen" windows don't actually take up the full monitor.

This is fine, but, get window position was hard coded to assume that it did.

I'm not sure why it was special cased instead of just returning the actual
position, but that seems to work. I need to test this on more platforms before
makign a PR, and verify that this assumption isn't made elsewhere.